+++ vtr-ui/src/app/beta/hardware-scan/widget/widget-hardware-scan/widget-hardware-scan.component.ts
@@ -9,11 +9,11 @@ import { HardwareScanService } from '../../services/hardware-scan/hardware-scan.
 })
 
 export class WidgetHardwareScanComponent implements OnInit {
-	@Input() title: string = '';
-	@Input() subTitle: string = '';
+	@Input() title: string = this.title || '';
+	@Input() subTitle: string = this.subTitle || '';
 	@Input() warningMessage = this.translate.instant('hardwareScan.warningMessage');
-	@Input() finalResultCode: string = '';
-	@Input() finalResultCodeText: string = '';
+	@Input() finalResultCode: string = this.finalResultCode || '';
+	@Input() finalResultCodeText: string = this.finalResultCodeText || '';
 	@Input() buttonText = '';
 	@Input() anchorText = '';
 	@Input() completeText = this.translate.instant('hardwareScan.complete');