+++ vtr-ui/src/app/components/modal/modal-gaming-legionedge/modal-gaming-legionedge.component.ts
@@ -1,7 +1,5 @@
 import { Component, OnInit, HostListener } from '@angular/core';
 import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';
-import { CommonService } from 'src/app/services/common/common.service';
-import { LocalStorageKey } from 'src/app/enums/local-storage-key.enum';
 
 @Component({
 	selector: 'vtr-modal-gaming-legionedge',
@@ -10,15 +8,7 @@ import { LocalStorageKey } from 'src/app/enums/local-storage-key.enum';
 })
 export class ModalGamingLegionedgeComponent implements OnInit {
 
-	liteGaming = false;
-	desktopMachine = false;
-	constructor(
-		public activeModal: NgbActiveModal,
-		private commonService: CommonService,
-	) { 
-		this.desktopMachine = this.commonService.getLocalStorageValue( LocalStorageKey.desktopType);
-		this.liteGaming = this.commonService.getLocalStorageValue( LocalStorageKey.liteGaming);
-	}
+	constructor(public activeModal: NgbActiveModal) { }
 
 	ngOnInit() {
 	}