+++ vtr-ui/src/app/components/pages/page-cpt/children/gaming/cptpage-autoclose/cptpage-autoclose.component.spec.ts
@@ -1,25 +1,25 @@
-// import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
-// import { CptpageAutocloseComponent } from './cptpage-autoclose.component';
+import { CptpageAutocloseComponent } from './cptpage-autoclose.component';
 
-// describe('CptpageAutocloseComponent', () => {
-//   let component: CptpageAutocloseComponent;
-//   let fixture: ComponentFixture<CptpageAutocloseComponent>;
+describe('CptpageAutocloseComponent', () => {
+  let component: CptpageAutocloseComponent;
+  let fixture: ComponentFixture<CptpageAutocloseComponent>;
 
-//   beforeEach(async(() => {
-//     TestBed.configureTestingModule({
-//       declarations: [ CptpageAutocloseComponent ]
-//     })
-//     .compileComponents();
-//   }));
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      declarations: [ CptpageAutocloseComponent ]
+    })
+    .compileComponents();
+  }));
 
-//   beforeEach(() => {
-//     fixture = TestBed.createComponent(CptpageAutocloseComponent);
-//     component = fixture.componentInstance;
-//     fixture.detectChanges();
-//   });
+  beforeEach(() => {
+    fixture = TestBed.createComponent(CptpageAutocloseComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
 
-//   it('should create', () => {
-//     expect(component).toBeTruthy();
-//   });
-// });
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});