+++ vtr-ui/src/app/components/pages/page-cpt/children/hardware-settings/cptpage-device-updates/cptpage-device-updates.component.spec.ts
@@ -1,25 +1,25 @@
-// import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
-// import { CptpageDeviceUpdatesComponent } from './cptpage-device-updates.component';
+import { CptpageDeviceUpdatesComponent } from './cptpage-device-updates.component';
 
-// describe('CptpageDeviceUpdatesComponent', () => {
-//   let component: CptpageDeviceUpdatesComponent;
-//   let fixture: ComponentFixture<CptpageDeviceUpdatesComponent>;
+describe('CptpageDeviceUpdatesComponent', () => {
+  let component: CptpageDeviceUpdatesComponent;
+  let fixture: ComponentFixture<CptpageDeviceUpdatesComponent>;
 
-//   beforeEach(async(() => {
-//     TestBed.configureTestingModule({
-//       declarations: [ CptpageDeviceUpdatesComponent ]
-//     })
-//     .compileComponents();
-//   }));
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      declarations: [ CptpageDeviceUpdatesComponent ]
+    })
+    .compileComponents();
+  }));
 
-//   beforeEach(() => {
-//     fixture = TestBed.createComponent(CptpageDeviceUpdatesComponent);
-//     component = fixture.componentInstance;
-//     fixture.detectChanges();
-//   });
+  beforeEach(() => {
+    fixture = TestBed.createComponent(CptpageDeviceUpdatesComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
 
-//   it('should create', () => {
-//     expect(component).toBeTruthy();
-//   });
-// });
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});