+++ vtr-ui/src/app/components/pages/page-cpt/children/hardware-settings/cptpage-security/cptpage-security.component.spec.ts
@@ -1,25 +1,25 @@
-// import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
-// import { CptpageSecurityComponent } from './cptpage-security.component';
+import { CptpageSecurityComponent } from './cptpage-security.component';
 
-// describe('CptpageSecurityComponent', () => {
-//   let component: CptpageSecurityComponent;
-//   let fixture: ComponentFixture<CptpageSecurityComponent>;
+describe('CptpageSecurityComponent', () => {
+  let component: CptpageSecurityComponent;
+  let fixture: ComponentFixture<CptpageSecurityComponent>;
 
-//   beforeEach(async(() => {
-//     TestBed.configureTestingModule({
-//       declarations: [ CptpageSecurityComponent ]
-//     })
-//     .compileComponents();
-//   }));
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      declarations: [ CptpageSecurityComponent ]
+    })
+    .compileComponents();
+  }));
 
-//   beforeEach(() => {
-//     fixture = TestBed.createComponent(CptpageSecurityComponent);
-//     component = fixture.componentInstance;
-//     fixture.detectChanges();
-//   });
+  beforeEach(() => {
+    fixture = TestBed.createComponent(CptpageSecurityComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
 
-//   it('should create', () => {
-//     expect(component).toBeTruthy();
-//   });
-// });
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});