+++ vtr-ui/src/app/components/pages/page-macrokey/page-macrokey.component.ts
@@ -75,6 +75,15 @@ export class PageMacrokeyComponent implements OnInit {
 			Page: 'macro-key'
 		};
 		this.cmsService.fetchCMSContent(queryOptions).subscribe((response: any) => {
+			const cardContentPositionF = this.cmsService.getOneCMSContent(
+				response,
+				'half-width-top-image-title-link',
+				'position-F'
+			)[0];
+			if (cardContentPositionF) {
+				this.cardContentPositionF = cardContentPositionF;
+			}
+
 			const cardContentPositionC = this.cmsService.getOneCMSContent(
 				response,
 				'half-width-title-description-link-image',
@@ -82,17 +91,8 @@ export class PageMacrokeyComponent implements OnInit {
 			)[0];
 			if (cardContentPositionC) {
 				this.cardContentPositionC = cardContentPositionC;
-			}
-
-			const cardContentPositionF = this.cmsService.getOneCMSContent(
-				response,
-				'inner-page-right-side-article-image-background',
-				'position-F'
-			)[0];
-			if (cardContentPositionF) {
-				this.cardContentPositionF = cardContentPositionF;
-				if (this.cardContentPositionF.BrandName) {
-					this.cardContentPositionF.BrandName = this.cardContentPositionF.BrandName.split('|')[0];
+				if (this.cardContentPositionC.BrandName) {
+					this.cardContentPositionC.BrandName = this.cardContentPositionC.BrandName.split('|')[0];
 				}
 			}
 		});