+++ vtr-ui/src/app/components/pages/page-privacy/feature/check-breached-accounts/confirmation/confirmation.component.ts
@@ -61,7 +61,7 @@ export class ConfirmationComponent implements OnInit, OnDestroy, AfterViewInit {
 			.pipe(
 				filter(Boolean),
 				takeUntil(instanceDestroyed(this))
-			).subscribe((userEmail: any) => {
+			).subscribe((userEmail) => {
 			this.userEmail = userEmail;
 		});
 	}