+++ vtr-ui/src/app/components/ui/ui-lighting-effect/ui-lighting-effect.component.ts
@@ -20,28 +20,19 @@ export class UiLightingEffectComponent implements OnInit, OnChanges {
 	@Output() public change = new EventEmitter<any>();
 	@Input() enableBrightCondition1: boolean;
 	@Input() showDescription: boolean;
-	@Input() isEffectChange:boolean = true;
-	@Input() showOptions:boolean = false;
+	public showOptions = false;
 	public buttonName: any = 'Show';
 	public selected = false;
 	public currentOption: string;
 	public currentDescription: string;
 	public selectedDescription: string;
 	@Input() effectOptionName: string;
-	public selectedOption: any = {
-		header: "",
-		id: "",
-		label: "",
-		metricitem: "",
-		name: "",
-		value: 1
-	};
+	public selectedOption: any;
 	@Input() defaultLang: any;
 	@ViewChild('dropdownLightingEle', { static: false })
 	dropdownEle: ElementRef;
 	intervalObj: any;
 	isItemsFocused = false;
-        @Output() public isEffectList = new EventEmitter<any>();
 	// for macrokey
 	@Input() public enableDescription = true;
 	@Input() isRecording = false;
@@ -76,7 +67,6 @@ export class UiLightingEffectComponent implements OnInit, OnChanges {
 		} else {
 			this.buttonName = 'Show';
 		}
-		this.isEffectList.emit(this.showOptions);
 	}
 
 	itemsFocused() {
@@ -102,10 +92,8 @@ export class UiLightingEffectComponent implements OnInit, OnChanges {
 
 	public optionSelected(option) {
 		this.selectedOption = option;
-		if(this.isEffectChange){
-			if (option.value === 4 || option.value === 8) {
-				this.effectOptionName = option.name;
-			}
+		if (option.value === 4 || option.value === 8) {
+			this.effectOptionName = option.name;
 		}
 
 		this.showOptions = false;
@@ -149,21 +137,5 @@ export class UiLightingEffectComponent implements OnInit, OnChanges {
 				}
 			}
 		}
-		// if (!isUndefined(changes.effectOptionName)) {
-		// 	if (changes.effectOptionName.previousValue !== changes.effectOptionName.currentValue) {
-		// 		this.effectOptionName = changes.effectOptionName.currentValue;
-		// 	}
-
-		// }
-		if(!isUndefined(changes.options)){
-			if(!isUndefined(changes.options.currentValue)){
-				if(!isUndefined(this.selectedValue)){
-					this.selectedOption = changes.options.currentValue.dropOptions.filter(
-						(option) => option.value === this.selectedValue
-					)[0];
-				}
-			}
-		}
-		console.log("this.selectedOption-----------------22222222222",this.selectedOption);
 	}
 }