+++ vtr-ui/src/app/components/widgets/widget-lighting/widget-lighting.component.ts
@@ -4,10 +4,8 @@ import { LocalStorageKey } from './../../../enums/local-storage-key.enum';
 import { CommonService } from 'src/app/services/common/common.service';
 import { GamingAllCapabilitiesService } from './../../../services/gaming/gaming-capabilities/gaming-all-capabilities.service';
 import { GamingLightingService } from './../../../services/gaming/lighting/gaming-lighting.service';
-import { Component, OnInit, Input, NgZone } from '@angular/core';
+import { Component, OnInit, Input } from '@angular/core';
 import { Router } from '@angular/router';
-import { VantageShellService } from 'src/app/services/vantage-shell/vantage-shell.service';
-import { EventTypes } from '@lenovo/tan-client-bridge';
 
 @Component({
 	selector: 'vtr-widget-lighting',
@@ -26,15 +24,10 @@ export class WidgetLightingComponent implements OnInit {
 	public isdriverpopup = false;
 	public isPopupVisible: any;
 	public defaultLanguage: any;
-	public ledlayoutversion:any;
-	public ledSwitchButtonFeature: boolean;
-	
 	constructor(
-		private ngZone: NgZone,
 		private gamingLightingService: GamingLightingService,
 		private commonService: CommonService,
-		private deviceService: DeviceService,
-		public shellServices: VantageShellService
+		private deviceService: DeviceService
 	) { }
 
 	ngOnInit() {
@@ -54,8 +47,6 @@ export class WidgetLightingComponent implements OnInit {
 	public getCapabilities() {
 		this.ledSetFeature = this.commonService.getLocalStorageValue(LocalStorageKey.ledSetFeature);
 		this.ledDriver = this.commonService.getLocalStorageValue(LocalStorageKey.ledDriver);
-		this.ledlayoutversion = this.commonService.getLocalStorageValue(LocalStorageKey.ledLayoutVersion);
-		this.ledSwitchButtonFeature = this.commonService.getLocalStorageValue(LocalStorageKey.LedSwitchButtonFeature);
 
 		if (this.ledSetFeature) {
 			if (LocalStorageKey.ProfileId !== undefined) {
@@ -74,9 +65,6 @@ export class WidgetLightingComponent implements OnInit {
 			this.isPopupVisible = true;
 		} else if (!this.ledSetFeature && !this.ledDriver) {
 			this.isLightingVisible = false;
-		}
-		if(this.ledSwitchButtonFeature){
-		 this.regLightingProfileIdChangeEvent();
 		}
 	}
 
@@ -135,26 +123,6 @@ export class WidgetLightingComponent implements OnInit {
 			}
 		} catch (error) {
 		}
-	}
-
-
-	public regLightingProfileIdChangeEvent(){
-		this.gamingLightingService.regLightingProfileIdChangeEvent();
-		this.shellServices.registerEvent(
-		  EventTypes.gamingLightingProfileIdChangeEvent,
-		  this.setProfileEvent.bind(this)
-		);
-	}
-
-	public setProfileEvent(profileId){
-		this.ngZone.run(()=>{
-			if(this.setprofId === profileId) return;
-			console.log("fn+space-------------home page-----------",profileId);
-			this.setprofId = profileId;
-			if (this.setprofId !== undefined) {
-				this.commonService.setLocalStorageValue(LocalStorageKey.ProfileId, this.setprofId);
-			}
-		})
 	}
 
 	public checkStatus(id) {